Extract ChatHistoryCard and update list
Add a new ChatHistoryCard component (ChatHistoryCard.tsx) that encapsulates the chat item UI and actions. Refactor ChatHistoryList.tsx to use ChatHistoryCard for each item, removing the duplicated markup and related imports/utils from the list. This reduces duplication and centralizes chat item rendering and action handling.
